clean: Organisation in Ordnerstruktur für das Frontend
