+++ frontend/src/App.js
@@ -355,178 +355,6 @@ function AppProvider({ children }) {
   );
 }
 
-function OrtForm() {
-  const { orte, addOrt, showNotification } = useContext(AppContext);
-  const [name, setName] = useState('');
-  const [adresse, setAdresse] = useState('');
-  const [istWohnort, setIstWohnort] = useState(false);
-  const [istDienstort, setIstDienstort] = useState(false);
-  const [istKirchspiel, setIstKirchspiel] = useState(false);
-  const [ortTyp, setOrtTyp] = useState('');
-  const hasDienstort = orte.some(ort => ort.ist_dienstort);
-  const hasWohnort = orte.some(ort => ort.ist_wohnort);
-  
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    addOrt({ name, adresse, istWohnort, istDienstort, istKirchspiel });
-    setName('');
-    setAdresse('');
-    setIstWohnort(false);
-    setIstDienstort(false);
-    setIstKirchspiel(false);
-    showNotification("Erfolg", "Der neue Ort wurde erfolgreich hinzugefügt.");
-  };
-  const sortedOrte = orte.sort((a, b) => a.name.localeCompare(b.name));
-  
-  return (
-    <div className="space-y-6">
-    <div className="card-container-highlight">
-    <h3 className="text-lg font-medium text-value mb-4">Ort hinzufügen</h3>
-    <p className="text-sm text-muted mb-6">
-    Hier können Sie neue Orte für Ihre Fahrten anlegen. Für die Struktur der Orteauswahl ist es Hilfreich den Orten Label zuzuordnen: Dienstort, Heimatort, Kirchspiel. Diese werden im Auswahlfeld weiter oben angezeigt.
-    </p>
-    
-    <form onSubmit={handleSubmit}>
-    <div className="flex flex-col sm:flex-row gap-4">
-    <div className="w-full sm:w-1/3">  {/* Breite angepasst */}
-    <label className="form-label">Name des Ortes</label>
-    <input
-    type="text"
-    value={name}
-    onChange={(e) => setName(e.target.value)}
-    className="form-input"
-    placeholder="z.B. Meldorf"
-    required
-    />
-    </div>
-    <div className="w-full sm:w-1/3">  {/* Breite angepasst */}
-    <label className="form-label">Adresse</label>
-    <input
-    type="text"
-    value={adresse}
-    onChange={(e) => setAdresse(e.target.value)}
-    className="form-input"
-    placeholder="Vollständige Adresse"
-    required
-    />
-    </div>
-    <div className="w-full sm:w-1/4">  {/* Breite angepasst */}
-    <label className="form-label">Art des Ortes</label>
-    <select
-    value={ortTyp}
-    onChange={(e) => {
-      const value = e.target.value;
-      setOrtTyp(value);
-      setIstWohnort(value === 'wohnort');
-      setIstDienstort(value === 'dienstort');
-      setIstKirchspiel(value === 'kirchspiel');
-    }}
-    className="form-select"
-    >
-    <option value="">Bitte wählen</option>
-    {!hasWohnort && <option value="wohnort">Wohnort</option>}
-    {!hasDienstort && <option value="dienstort">Dienstort</option>}
-    <option value="kirchspiel">Kirchspiel</option>
-    <option value="none">Sonstiger Ort</option>
-    </select>
-    </div>
-    <div className="flex items-end w-full sm:w-auto">
-    <button type="submit" className="btn-primary w-full sm:w-auto">
-    Hinzufügen
-    </button>
-    </div>
-    </div>
-    </form>
-    </div>
-    </div>
-  );
-}
-
-function DistanzForm() {
-  const { orte, addDistanz, distanzen, showNotification } = useContext(AppContext);
-  const [vonOrtId, setVonOrtId] = useState('');
-  const [nachOrtId, setNachOrtId] = useState('');
-  const [distanz, setDistanz] = useState('');
-  const [existingDistanz, setExistingDistanz] = useState(null);
-  
-  useEffect(() => {
-    if (vonOrtId && nachOrtId) {
-      const existing = distanzen.find(d => 
-        (d.von_ort_id === parseInt(vonOrtId) && d.nach_ort_id === parseInt(nachOrtId)) ||
-        (d.von_ort_id === parseInt(nachOrtId) && d.nach_ort_id === parseInt(vonOrtId))
-      );
-      setExistingDistanz(existing);
-      if (existing) {
-        setDistanz(existing.distanz.toString());
-      } else {
-        setDistanz('');
-      }
-    } else {
-      setExistingDistanz(null);
-      setDistanz('');
-    }
-  }, [vonOrtId, nachOrtId, distanzen]);
-  
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    addDistanz({ vonOrtId, nachOrtId, distanz: parseInt(distanz) });
-    setVonOrtId('');
-    setNachOrtId('');
-    setDistanz('');
-    setExistingDistanz(null);
-    showNotification("Erfolg", "Die neue Distanz wurde erfolgreich hinzugefügt.");
-  };
-  
-  const sortedOrte = orte.sort((a, b) => a.name.localeCompare(b.name));
-  
-  return (
-    <div className="mb-4">
-    <div className="card-container-highlight">
-    <form onSubmit={handleSubmit} className="flex flex-col sm:flex-row gap-4">
-    <div className="w-full sm:flex-1">
-    <select
-    value={vonOrtId}
-    onChange={(e) => setVonOrtId(e.target.value)}
-    className="form-select"
-    required
-    >
-    <option value="">Von Ort auswählen</option>
-    {renderOrteOptions(orte)}
-    </select>
-    </div>
-    
-    <div className="w-full sm:flex-1">
-    <select
-    value={nachOrtId}
-    onChange={(e) => setNachOrtId(e.target.value)}
-    className="form-select"
-    required
-    >
-    <option value="">Nach Ort auswählen</option>
-    {renderOrteOptions(orte)}
-    </select>
-    </div>
-    
-    <div className="w-full sm:w-36">
-    <input
-    type="number"
-    value={distanz}
-    onChange={(e) => setDistanz(e.target.value)}
-    placeholder="km"
-    className="form-input"
-    required
-    />
-    </div>
-    
-    <button type="submit" className="btn-primary">
-    {existingDistanz ? 'Aktualisieren' : 'Hinzufügen'}
-    </button>
-    </form>
-    </div>
-    </div>
-  );
-}
-
 function FahrtenListe() {
   const { fahrten, selectedMonth, setSelectedMonth, fetchFahrten, deleteFahrt, updateFahrt, orte, fetchMonthlyData, showNotification, summary, setFahrten } = useContext(AppContext);
   const [expandedFahrten, setExpandedFahrten] = useState({});
@@ -2312,362 +2140,6 @@ function MonthlyOverview() {
     </div>
     </div>
     </Modal>
-    </div>
-  );
-}
-
-function OrteListe() {
-  const { orte, updateOrt, deleteOrt, showNotification } = useContext(AppContext);
-  const [editingOrt, setEditingOrt] = useState(null);
-  const [isCollapsed, setIsCollapsed] = useState(false);
-  const [sortConfig, setSortConfig] = useState({ key: 'name', direction: 'ascending' });
-
-  const handleEdit = (ort) => {
-    setEditingOrt({ ...ort });
-  };
-  
-  const handleSave = () => {
-    const updatedOrt = {
-      ...editingOrt,
-      ist_wohnort: editingOrt.ist_wohnort !== undefined ? editingOrt.ist_wohnort : false,
-      ist_dienstort: editingOrt.ist_dienstort !== undefined ? editingOrt.ist_dienstort : false,
-      ist_kirchspiel: editingOrt.ist_kirchspiel !== undefined ? editingOrt.ist_kirchspiel : false
-    };
-    updateOrt(editingOrt.id, updatedOrt);
-    setEditingOrt(null);
-    showNotification("Erfolg", "Der Ort wurde erfolgreich aktualisiert.");
-  };
-  
-  
-  
-  const handleDelete = async (id) => {
-    showNotification(
-      "Ort löschen",
-      "Sind Sie sicher, dass Sie diesen Ort löschen möchten?",
-      async () => {
-        try {
-          await deleteOrt(id);
-          showNotification("Erfolg", "Der Ort wurde erfolgreich gelöscht.");
-        } catch (error) {
-          console.error('Fehler beim Löschen des Ortes:', error);
-          showNotification("Fehler", "Dieser Ort kann nicht gelöscht werden, da er in Fahrten verwendet wird.");
-        }
-      },
-      true // showCancel
-    );
-  };
-  
-  const sortedOrte = React.useMemo(() => {
-    let sortableOrte = [...orte];
-    if (sortConfig.key !== null) {
-      sortableOrte.sort((a, b) => {
-        if (a[sortConfig.key] < b[sortConfig.key]) {
-          return sortConfig.direction === 'ascending' ? -1 : 1;
-        }
-        if (a[sortConfig.key] > b[sortConfig.key]) {
-          return sortConfig.direction === 'ascending' ? 1 : -1;
-        }
-        return 0;
-      });
-    }
-    return sortableOrte;
-  }, [orte, sortConfig]);
-  
-  const requestSort = (key) => {
-    let direction = 'ascending';
-    if (sortConfig.key === key && sortConfig.direction === 'ascending') {
-      direction = 'descending';
-    }
-    setSortConfig({ key, direction });
-  };
-  
-  const getOrtStatus = (ort) => {
-    if (ort.ist_wohnort) return 'wohnort';
-    if (ort.ist_dienstort) return 'dienstort';
-    if (ort.ist_kirchspiel) return 'kirchspiel';
-    return '';
-  };
-  
-  const getOrtStatusLabel = (ort) => {
-    if (ort.ist_wohnort) return 'Wohnort';
-    if (ort.ist_dienstort) return 'Dienstort';
-    if (ort.ist_kirchspiel) return 'Kirchspiel';
-    return 'Sonstiger Ort';
-  };
-  
-  const handleStatusChange = (e) => {
-    const value = e.target.value;
-    setEditingOrt({
-      ...editingOrt,
-      ist_wohnort: value === 'wohnort',
-      ist_dienstort: value === 'dienstort',
-      ist_kirchspiel: value === 'kirchspiel'
-    });
-  };
-  
-  return (
-    <div className="table-container">
-    <table className="w-full">
-    <thead>
-    <tr className="table-head-row">
-    <th className="table-header" onClick={() => requestSort('name')}>
-    Name {sortConfig.key === 'name' && (
-      <span className="text-muted">{sortConfig.direction === 'ascending' ? '↑' : '↓'}</span>
-    )}
-    </th>
-    <th className="table-header-sm" onClick={() => requestSort('adresse')}>
-    Adresse {sortConfig.key === 'adresse' && (
-      <span className="text-muted">{sortConfig.direction === 'ascending' ? '↑' : '↓'}</span>
-    )}
-    </th>
-    <th className="table-header">Status</th>
-    <th className="table-header text-right">Aktionen</th>
-    </tr>
-    </thead>
-    <tbody className="divide-y divide-primary-50 dark:divide-primary-700">
-    {sortedOrte.map((ort) => (
-      <tr key={ort.id} className="table-row">
-      <td className="table-cell">
-      {editingOrt?.id === ort.id ? (
-        <input
-        value={editingOrt.name}
-        onChange={(e) => setEditingOrt({ ...editingOrt, name: e.target.value })}
-        className="form-input"
-        />
-      ) : (
-        <div className="flex flex-col">
-        <span className="text-value">{ort.name}</span>
-        <span className="text-muted text-xs sm:hidden">
-        {ort.adresse}
-        </span>
-        </div>
-      )}
-      </td>
-      <td className="table-cell hidden md:table-cell">
-      {editingOrt?.id === ort.id ? (
-        <input
-        value={editingOrt.adresse}
-        onChange={(e) => setEditingOrt({ ...editingOrt, adresse: e.target.value })}
-        className="form-input"
-        />
-      ) : (
-        <span className="text-value">{ort.adresse}</span>
-      )}
-      </td>
-      <td className="table-cell">
-      {editingOrt?.id === ort.id ? (
-        <select
-        value={getOrtStatus(editingOrt)}
-        onChange={handleStatusChange}
-        className="form-select"
-        >
-        <option value="">Sonstiger Ort</option>
-        <option value="wohnort">Wohnort</option>
-        <option value="dienstort">Dienstort</option>
-        <option value="kirchspiel">Kirchspiel</option>
-        </select>
-      ) : (
-        <span className="text-value">{getOrtStatusLabel(ort)}</span>
-      )}
-      </td>
-      <td className="table-cell">
-      <div className="flex justify-end gap-2">
-      {editingOrt?.id === ort.id ? (
-        <button
-        onClick={handleSave}
-        className="table-action-button-primary"
-        title="Speichern"
-        >
-        ✓
-        </button>
-      ) : (
-        <>
-        <button
-        onClick={() => handleEdit(ort)}
-        className="table-action-button-primary"
-        title="Bearbeiten"
-        >
-        ✎
-        </button>
-        <button
-        onClick={() => handleDelete(ort.id)}
-        className="table-action-button-secondary"
-        title="Löschen"
-        >
-        ×
-        </button>
-        </>
-      )}
-      </div>
-      </td>
-      </tr>
-    ))}
-    </tbody>
-    </table>
-    </div>
-  );
-}
-
-function DistanzenListe() {
-  const { distanzen, orte, updateDistanz, deleteDistanz, showNotification } = useContext(AppContext);
-  const [editingDistanz, setEditingDistanz] = useState(null);
-  const [isCollapsed, setIsCollapsed] = useState(false);
-  const [sortConfig, setSortConfig] = useState({ key: 'von_ort_id', direction: 'ascending' });
-  
-  const handleEdit = (distanz) => {
-    setEditingDistanz({ ...distanz });
-  };
-  
-  const handleSave = () => {
-    updateDistanz(editingDistanz.id, editingDistanz);
-    setEditingDistanz(null);
-    showNotification("Erfolg", "Die Distanz wurde erfolgreich aktualisiert.");
-  };
-  
-  const getOrtName = (id) => {
-    const ort = orte.find(o => o.id === id);
-    return ort ? ort.name : 'Unbekannt';
-  };
-  
-  const handleDelete = async (id) => {
-    showNotification(
-      "Distanz löschen",
-      "Sind Sie sicher, dass Sie diese Distanz löschen möchten?",
-      async () => {
-        try {
-          await deleteDistanz(id);
-          showNotification("Erfolg", "Die Distanz wurde erfolgreich gelöscht.");
-        } catch (error) {
-          console.error('Fehler beim Löschen der Distanz:', error);
-          showNotification("Fehler", "Beim Löschen der Distanz ist ein Fehler aufgetreten.");
-        }
-      },
-      true // showCancel
-    );
-  };
-  
-  const sortedDistanzen = useMemo(() => {
-    let sortableDistanzen = [...distanzen];
-    if (sortConfig.key !== null) {
-      sortableDistanzen.sort((a, b) => {
-        if (sortConfig.key === 'von_ort_id' || sortConfig.key === 'nach_ort_id') {
-          const ortA = orte.find(o => o.id === a[sortConfig.key]);
-          const ortB = orte.find(o => o.id === b[sortConfig.key]);
-          return sortConfig.direction === 'ascending' 
-          ? ortA.name.localeCompare(ortB.name)
-          : ortB.name.localeCompare(ortA.name);
-        } else {
-          if (a[sortConfig.key] < b[sortConfig.key]) {
-            return sortConfig.direction === 'ascending' ? -1 : 1;
-          }
-          if (a[sortConfig.key] > b[sortConfig.key]) {
-            return sortConfig.direction === 'ascending' ? 1 : -1;
-          }
-          return 0;
-        }
-      });
-    }
-    return sortableDistanzen;
-  }, [distanzen, sortConfig, orte]);
-  
-  const requestSort = (key) => {
-    let direction = 'ascending';
-    if (sortConfig.key === key && sortConfig.direction === 'ascending') {
-      direction = 'descending';
-    }
-    setSortConfig({ key, direction });
-  };
-  
-  return (
-    <div className="table-container">
-    <table className="w-full">
-    <thead>
-    <tr className="table-head-row">
-    <th className="table-header" onClick={() => requestSort('von_ort_id')}>
-    Von {sortConfig.key === 'von_ort_id' && (
-      <span className="text-muted">{sortConfig.direction === 'ascending' ? '↑' : '↓'}</span>
-    )}
-    </th>
-    <th className="table-header-sm" onClick={() => requestSort('nach_ort_id')}>
-    Nach {sortConfig.key === 'nach_ort_id' && (
-      <span className="text-muted">{sortConfig.direction === 'ascending' ? '↑' : '↓'}</span>
-    )}
-    </th>
-    <th className="table-header" onClick={() => requestSort('distanz')}>
-    <span className="sm:hidden">km</span>
-    <span className="hidden sm:inline">Distanz (km)</span>
-    {sortConfig.key === 'distanz' && (
-      <span className="text-muted">{sortConfig.direction === 'ascending' ? '↑' : '↓'}</span>
-    )}
-    </th>
-    <th className="table-header text-right">Aktionen</th>
-    </tr>
-    </thead>
-    <tbody className="divide-y divide-primary-50 dark:divide-primary-700">
-    {sortedDistanzen.map((distanz) => (
-      <tr key={distanz.id} className="table-row">
-      <td className="table-cell">
-      <div className="flex flex-col">
-      <span className="text-value">{getOrtName(distanz.von_ort_id)}</span>
-      <span className="text-muted text-xs sm:hidden">
-      → {getOrtName(distanz.nach_ort_id)}
-      </span>
-      </div>
-      </td>
-      <td className="table-cell hidden md:table-cell">
-      <span className="text-value">{getOrtName(distanz.nach_ort_id)}</span>
-      </td>
-      <td className="table-cell text-right">
-      {editingDistanz?.id === distanz.id ? (
-        <input
-        type="number"
-        value={editingDistanz.distanz}
-        onChange={(e) =>
-          setEditingDistanz({
-            ...editingDistanz,
-            distanz: parseInt(e.target.value),
-          })
-        }
-        className="form-input w-16"
-        />
-      ) : (
-        <span className="text-value">{distanz.distanz}</span>
-      )}
-      </td>
-      <td className="table-cell">
-      <div className="flex justify-end gap-2">
-      {editingDistanz?.id === distanz.id ? (
-        <button
-        onClick={handleSave}
-        className="table-action-button-primary"
-        title="Speichern"
-        >
-        ✓
-        </button>
-      ) : (
-        <>
-        <button
-        onClick={() => handleEdit(distanz)}
-        className="table-action-button-primary"
-        title="Bearbeiten"
-        >
-        ✎
-        </button>
-        <button
-        onClick={() => handleDelete(distanz.id)}
-        className="table-action-button-secondary"
-        title="Löschen"
-        >
-        ×
-        </button>
-        </>
-      )}
-      </div>
-      </td>
-      </tr>
-    ))}
-    </tbody>
-    </table>
     </div>
   );
 }
